Create mastermind_tester
This is a slightly modified version of the program designed to test reliability over a large number of guess attempts. I can't take all the credit for the idea, as Wai Lee recommended that I write something like this. I didn't really have a high level of confidence in the ability for random guesses to get under 10 consistently until I saw it succeed 1000 times in a row.