Update mastermind_tester.rb
There was an artifact of the an old version of the program that got around the destructive nature of the testing process by cloning the set and pulling new guesses from that. That approach never really worked well, so I switched to rebuilding the set at the end of every loop. The "set_clone" variable snuck into the code in the line that picks the new guesses. Silly mistake, corrected now.